+++ src/nestjs/openapi/index.ts
@@ -1,1 +1,4 @@
 export { ApiConstantProperty } from './api-constant-property.decorator.js';
+export { ApiErrorDescription } from './api-error-description.decorator.js';
+export { ApiErrorResponses } from './api-error-responses.decorator.js';
+export { ApiErrorStatusCode } from './api-error-status-code.decorator.js';